+++ app/features/products/productImages.ts
@@ -169,6 +169,8 @@ import { たかねこフェスvol6_ミニフォト } from "./2026/2026-07-12_ミ
 import { たかねこフェスvol6_生写真 } from "./2026/2026-07-12_生写真「たかねこフェス vol.6」";
 import { TIF2026限定ライブフォト1_ミニフォト } from "./2026/2026-07-31_ミニフォトカード「TIF2026 限定ライブフォト 1」";
 import { TIF2026限定ライブフォト2_ミニフォト } from "./2026/2026-07-31_ミニフォトカード「TIF2026 限定ライブフォト 2」";
+import { STELLA_FC抽選会_アンブレラマーカー } from "./2026/2026-08-06_アンブレラマーカー「4th Anniversary Special LIVE 〜STELLA〜 FC抽選会」";
+import { STELLA_FC抽選会_自撮りステッカー } from "./2026/2026-08-06_ステッカー「4th Anniversary Special LIVE 〜STELLA〜 FC抽選会」";
 import { ItemDescription, ProductLine, RandomGoods, TradeTextType } from "./product";
 
 export const TAKANEKO_PHOTOS_FEATURED: RandomGoods[] = [
@@ -179,6 +181,8 @@ export const TAKANEKO_PHOTOS_FEATURED: RandomGoods[] = [
 ];
 
 export const TAKANEKO_PHOTOS: RandomGoods[] = [
+  STELLA_FC抽選会_アンブレラマーカー,
+  STELLA_FC抽選会_自撮りステッカー,
   TIF2026限定ライブフォト2_ミニフォト,
   TIF2026限定ライブフォト1_ミニフォト,
   たかねこフェスvol6_生写真,